+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Other:            GnssSinexBias2SignalBias: Transform combined X signals to actually transmitted signals.
 - Other:            GnssParametrizationAmbiguities: Consider more resolved ambiguities in the next processing steps.
 - Other:            GriddedData(TimeSeries)2PotentialCoefficients: Reduced memory consumption in parallel mode.
+- Other:            dataVariables/statistics: ignore NaN entries.
 
 # Release 2025-11-15
 - New program:      StarCamera2Orbex (replaces GnssAttitude2Orbex).